examples/etcd_app/requirements.txt
python-etcd>=0.4.3
vyper
